+++ src/main/java/com/manchui/domain/review/service/ReviewService.java
@@ -1,9 +1,9 @@
-package com.manchui.domain.service;
+package com.manchui.domain.review.service;
 
-import com.manchui.domain.dto.review.ReviewCreateRequest;
-import com.manchui.domain.dto.review.ReviewCreateResponse;
-import com.manchui.domain.dto.review.ReviewDetailPagingResponse;
-import com.manchui.domain.dto.review.ReviewScoreResponse;
+import com.manchui.domain.review.dto.ReviewCreateRequest;
+import com.manchui.domain.review.dto.ReviewCreateResponse;
+import com.manchui.domain.review.dto.ReviewDetailPagingResponse;
+import com.manchui.domain.review.dto.ReviewScoreResponse;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
